requirements.txt
certifi==2025.1.31
Flask==3.1.0
flask_cors==5.0.1
numpy==2.2.3
Pillow==10.0.0
Pillow==11.1.0
protobuf==6.30.0
pymongo==4.11.2
Requests==2.32.3
streamlit==1.35.0
tensorflow==2.18.0
Werkzeug==3.1.3
